feat : PurchaseAmount -> PurchaseInfomation 객체로 수정
